+++ changelog.md
@@ -6,6 +6,7 @@
 - Redizajnovaná prvá stránka FW, teraz prístupná bez prihlásenia
 - Vytvorený kontroler `AdminController`, ktorý je dostupný až po prihlásení
 - Fix, ak sa DB tabuľka alebo stĺpec volá rovnako ako rezervované slovo v DB
+- Added `files` attribute to `Request` class
 
 ### Version 1.0 
 - Úvodná verzia FW